requirements.txt
apache-airflow[celery,postgres,redis]==2.8.0
supabase==2.3.4
polars==0.20.0
duckdb==0.10.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9